allow to set required accuracy with one more digit precision
Now one can set fractional percent values for the required accuracy
to unlock the next lesson. Sine the difference between 97% and 98%
accuracy actually means one may make only two thirds of the errors
to pass a more fine granual setting is really helpful.

FEATURE: 283591
FIXED-IN: 2.0

+++ src/declarativeitems/preferencesproxy.cpp
@@ -76,12 +76,12 @@ void PreferencesProxy::setRequiredStrokesPerMinute(int strokesPerMinute)
     Preferences::setRequiredStrokesPerMinute(strokesPerMinute);
 }
 
-int PreferencesProxy::requiredAccuracy() const
+double PreferencesProxy::requiredAccuracy() const
 {
     return Preferences::requiredAccuracy();
 }
 
-void PreferencesProxy::setRequiredAccuracy(int accuracy)
+void PreferencesProxy::setRequiredAccuracy(double accuracy)
 {
     Preferences::setRequiredStrokesPerMinute(accuracy);
 }
